chore: download browsers out of process
